requirements.txt
langchain==0.2.1
langchain-community
langchain-core==0.2.1
chromadb==0.5.0
sentence-transformers==2.7.0
huggingface-hub==0.23.0
openai==1.30.1  # Optional alternative
pypdf
unstructured==0.13.4
streamlit==1.35.0  # For simple UI
python-dotenv==1.0.1
jupyter==1.0.0  # For experimentation
ipython==8.22.0  # Enhanced REPL